Sphere and ray intersection logic added. It should be further thested currently parameters should give two positive t values but give one+ one-

+++ src/models/Material.cpp
@@ -1,9 +1,9 @@
 #include "Material.h"
 #include <vector>
 
-Material::Material(MaterialType type, int ambient[3], int diffuse[3], int specular[3]){
+Material::Material(MaterialType type, float ambient[3], float diffuse[3], float specular[3]){
     this->materialType = type;
-    this->ambientProp = std::vector<int>(ambient, ambient+3);
-    this->diffuseProp = std::vector<int>(diffuse, diffuse+3);
-    this->specularProp = std::vector<int>(specular, specular+3);
+    this->ambientProp = std::vector<float>(ambient, ambient+3);
+    this->diffuseProp = std::vector<float>(diffuse, diffuse+3);
+    this->specularProp = std::vector<float>(specular, specular+3);
 }
